Add files via upload
This was presented in Marikina Metropolitan Manila on 21 October 2025 for the Student Summit held by IECEP National.  The aim was to inform students that the present way the subjects are taught have a reason but due to its structured design and non major subjects its not enough to teach product development. A way forward is to instead adopt a UX approach to making needs research, design planning, execution and testing so that its an R&D approach not a Research Approach. This can make them better prepared for the industry needs rather than just the academic demand.  Inserted into this presentation is the difference of Research and R&D, Engineering Design Approach based on Molvani, Issues with Engineering Design Approach in Academic Environments, Design Thinking Approach, Design Thinking Approach with UX integrated.  This is the original material by Voltaire Dupo. The information was his to collect and analyse.